Replace custom integer types

+++ src-headers/defs.h
@@ -164,7 +164,7 @@ void swtch(context_t **, context_t *);
 
 // spinlock.c
 void acquire(struct spinlock *);
-void getcallerpcs(void *, uint *);
+void getcallerpcs(void *, uintptr_t *);
 int holding(struct spinlock *);
 void initlock(struct spinlock *, char *);
 void release(struct spinlock *);
